Fix slice numbering in `sct_analyze_lesion` QC and CLI output (#4597)
## Description

sct_analyze_lesion reorients the image to RPI for consistent processing. Because
of this, the slice numbers in the QC and excel spreadsheet correspond to the RPI
slice numbers. But if the image is e.g. LAS instead, there is a LR flip, meaning that
the RPI (right -> left) slice numbers do not correspond to the slice numbers in the
original orientation. So, if the image is opened in, say, FSLeyes, the slice numbers
will not correspond.

This change ensures that we display the original orientation slice numbers in the
QC report and spreadsheet.

## Linked issues

Resolves
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4592

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>