Skip `-h` duration test on macOS CI runners due to sporadic runtimes (#4605)
The problem described in #4515 turned out to not be a "one in a blue
moon" kind of slowdown. I'd figure it's more a 50/50 chance to get a
slow macOS runner, and so I've been having to repeatedly restart the
cron jobs to ensure passing runs.

We could go with a fancier solution (checking the runner hardware,
having a different "passing" time for macOS, etc.) but given that our
other platforms (Ubuntu, Windows) have _never_ failed this check, I
think it should be OK to rely on them as a signal for whether a PR is
adding a slow import.

Fixes #4515.

(cherry picked from commit 9fa6961d8b1667f7ac8977a64be9f78637e0559b)