Add more intuitive feedback when no metrics can be extracted from input data (#4606)
## Description

In #4552, jcohenadad tried extracting metrics from a user's data, but
the template/atlas files were empty. This meant that the specified
vertlevels had no corresponding slices, and thus no metrics could be
extracted.

We already have a conditional for when this occurs.
Yet, we don't provide any feedback to the user when this happening. Even
worse, when _all_ of the slicegroups are empty, the `save_as_csv`
function fails in an ugly way.

So, this PR ensures that warnings are thrown + no crash occurs, making
it easier for the user to debug their own data.

## Linked issues

Fixes #4552.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>