`test_cli_sct_dmri_moco`: Update mask and results for `-m` test
Motion correction is highly sensitive to the mask size/shape/location, so the motion
parameters were bound to change.

I also increased the mask size and altered the position, because now that we're binarizing
the mask at 0.5, we need a much larger mask to cover the same region (due to the use of a
gaussian mask).