Allow error code 406 in broken link checker (#4610)
Our "broken link checker" CI workflow started failing a few days ago due
to the Radiopaedia.org domain. The links open fine in browser, but they
can't be fetched by curl/wget. I also tried using https://http.app/, but
got the same error.

I spent an hour or so deep-diving into the error code, but I couldn't
identify the source of the error, so my best guess is that this is a
configuration issue in Radiopaedia.org's CloudFlare settings, or an
attempt to stop crawlers from scraping Radiopaedia's articles (due to
recent LLM controversies re: ignoring `robots.txt`).

We could have also filtered the Radiopaedia domain (as we had done in
the past with other domains), but due to the limited scope of the 406
error code, I figure that filtering it would allow us to continue
checking Radiopaedia for other return codes (in case their config
changes again in the future).