`sct_get_centerline`: Fetch `arr` from reoriented `im_centerline`
This is a lot easier than trying to reorient `arr_centerline` inside `get_centerline`, since
that would potentially involve a new parameter, some complicated math, and the risk of
affecting other CLI scripts that call `get_centerline`.

However, it does mean that the effect of the `space` parameter is no longer considered. So,
we will need to use a different approach to convert the pixel coordinates to physical ones.