`centerline/core.py`: Fix math error when generating `space=phys`
In order to generate accurate physical (world) coordinates, the affine matrix must map the
voxel coordinates to RAS+/LPI- space. This is true by default for all NIFTI images generated
using programs such as dcm2niix, and it remains true even if the image is reoriented at a
later time. This is because, even if the data array is reoriented, the affine will be updated
in tandem (either by transposing the affine's columns, or by flipping axes through
multiplying affines by -1).

The problem is, the coordinates must be multiplied *by the affine corresponding to orientation
of those coordinates*. And in this case, while `im_seg` was reoriented back to the native
orientation, the coordinates are actually in RPI! Meaning, the matmul produced invalid
coordinates that did not match the readings given by e.g. FSLeyes.

We fix this issue by skipping the reorientation of `im_seg` back to the original orientation.
This means that when `im_seg` is used for its `transfo_pix2phys` method, we'll get RPI affine
* RPI coordinates = RAS+/LPI- world coordinates. This reorientation isn't necessary because
we make a copy of `im_seg` (to avoid mutating the input). But, in previous iterations,
`im_seg` was actually mutated, hence why there used to be a need for reorientation.

----

Note: Even in older iterations of SCT, this orientation mismatch existed.

> tl;dr, there used to be a secondary, confusingly-named "_get_centerline()" function whose
> entire purpose was to generate a centerline from physical coordinates. All usages of
> `space="phys"` stem from the former usage of `_get_centerline()`, so their logic is
> identical in all cases. (Refer to #4010 for context of the changes)
>
> Also, in that function, we used to call `img.transfo_pix2phys` (original orientation) on
> `arr_ctl`, whose coordinates were in RPI orientation. Meaning, the refactor in #4010 only
> propagated the existing buggy `pix2phys` call.

----

Even if the old logic was seemingly buggy, we should look at how the output of `space="phys"`
(i.e. the old calls to `_get_centerline()`) actually gets used, so that we can see the
consequences of the switch to RAS+/LPI- world coordinates.

In all cases, the sole usage of the `arr_ctl` and `arr_ctl_dev` variables (when space="phys")
is to pass them to the constructor of `types.Centerline`. And, `Centerline.__init__` processes
the coordinates in a very complex way; it does not contain a notion of voxel orientation, but
instead works entirely in physical space, with notions of "progressive length" and
"incremental length". So, it is difficult for me to grasp the consequences of this change.

Note that the scripts affected by the change appear to be:
- sct_straighten_spinalcord
- sct_process_segmentation -pmj
- sct_compute_compression

The question I have now is: Were the "length" attributes incorrect before? Or have I made them
incorrect by fixing the output of `pix2phys` to be RAS+/LPI-? I will let the automated testing
run, and see how this change affects the above scripts.