Automatically format `"\n"` into ``"""` (multiline strings)
Multiline strings play MUCH nicer with the argparse formatter. With single strings, the
precise placement of each `\n` would affect whether a newline was created in the output,
and maintaining them would be much harder in the long-run.

Note: This output was generated by a formatting script that I wrote.