`shell.py`: Replace broken link to removed project
Huzzah! The broken link checker caught an actually broken link, as the project we were
referencing has since been deleted from both RTD and PyPI. It looks like we borrowed
code from someone's personal project that they decided to scrub from the internet.

I've taken the time to identify the source file, highlight the relevant lines, then
archive the associated link. However, because archive.is *also* throws an error when
run through curl (429 - Too Many Requests), I've omitted the "http" part of the link
to keep the link checker from failing again.

Also, I'm committing this directly to master because it is a small link change in a
non-user-facing docstring, and I figure that it is similar enough in spirit to me
committing to master to update SCT's citation page that committing to master is OK.