`check-url.sh`: Use `-I` to only send a HEAD request
Oh no! It looks like we were accidentally being bad netizens. I'm not sure why we send
a HEAD request to get `full_info` and `LOCATION`, but send a full GET request for the
others. This is... incredibly wasteful, and a big nuisance towards the sites.

Anecdotally, when sending a HEAD request to arxiv, I was actually unable to trigger a
429 request no matter how hard I spammed. But, when I excluded `-I`, things suddenly
slowed to a crawl. So, I think adding `-I` make a huge difference here.