`check-url.sh`: Re-use common curl args via a shell array
Note 1: I've added comments + changed -I to --head for clarity's sake.

Note 2: It's recommended to use an array for building a command line.

I ran into this because I first tried to use a string, but then ran into shellcheck's
double-quoting warning. Because I *wanted* to split on spaces, I knew this was an
exception. And, sure enough, the following appears in the "Exceptions" section for
https://www.shellcheck.net/wiki/SC2086:

> Sometimes you want to split on spaces, like when building a command line. Just
> quoting this doesn't work. Instead, you should have used an array.