`download.py`: Ask for permission before deleting `-o` directory
I've followed the logical suggestions from @mguaypaq in https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4562#issuecomment-2229691874, however
I've added an extra check for non-emptiness, since an empty directory seems OK to
remove by default (seeing as it's just going to get remade anyway).

Note that my changes only provide protection for callers of `install_named_dataset`,
meaning that `install_data` by itself is not protected. But, `install_named_dataset` is
the only place we do the "default vs. non-default location" check, so I think it's the
only place we *can* add this safety valve.

The only other usage of `install_data` is `install_model` (`sct_deepseg`), however,
and sct_deepseg does not allow specifying a custom output directory when installing.
So, I think this is suitable? However, the fact that `install_data` has `keep=False`
by default is a little worrisome, as it could lead to a footgun if we ever use the
function anywhere else.

That said, I imagine that if we ever do re-use `install_data` elsewhere, we'll be
especially cognizant of the risk as maintainers given the prominance of the `keep` arg.