`math.py`: Add point label method to `dilate()`
Since 99.99% of voxels are 0, we don't *need* to run dilation on every single voxel, which is what skimage will do by default.

Instead, we can fetch the nonzero voxels, extract an area the size of the footprint, run dilation on that area, and then put it back.

I was concerned that this might behave differently in cases where we have two labels very close together. But, I think it ends up working similarly, with the larger label dominating the smaller label (thanks to the use of `maximum` internally).

There's an even faster approach (just multiplying `footprint` by the label value, then setting the area equal to this), but I think that would actually cause overwriting for nearby labels, and dilation is fast enough here.