`inference.py`: Refactor multiclass image saving code
I think part of the reason this issue happened is that `targets` isn't used in the scope of
`segment_nnunet`. So, to make `targets` matter just a little bit more, I've deduplicated the
image saving code, and harmonized the variables between the rootlets case and the non-rootlets
case.

NB: The code I've changed is actually just the tmpdir-saving code. (The image will be saved once
more in the `sct_deepseg` context using the `target` value.) So, this refactor is of little
functional consequence, beyond aiding debugging (since the tmpdir filename should now match the
outdir filename).