Update contrast agnostic model to r20241024 (improved for SCI and whole-spine T1/T2 images) (#4661)
<!-- Hi, and thank you for submitting a Pull Request! The checklist
below is a brief summary of steps found in the NeuroPoly Contributing
Guidelines, which can be found here:
https://intranet.neuro.polymtl.ca/onboarding/software-development.html#software-development.
-->

## Checklist

#### GitHub

- [x] I've given this PR a concise, self-descriptive, and meaningful
title
- [ ] I've linked relevant issues in the PR body
- [x] I've applied [the relevant
labels](https://intranet.neuro.polymtl.ca/onboarding/software-development.html#pr-labels)
to this PR
- [ ] I've applied a [release
milestone](https://github.com/spinalcordtoolbox/spinalcordtoolbox/milestones)
(major, minor, patch) in line with [Semantic Versioning
guidelines](https://github.com/spinalcordtoolbox/spinalcordtoolbox/wiki/Misc%3A-Creating-a-new-release#convention-for-naming-releases)
- [x] I've assigned a reviewer

<!-- For the title, please observe the following rules:
	- Provide a concise and self-descriptive title
- Do not include the applicable issue number in the title, do it in the
PR body
	- If the PR is not ready for review, convert it to a draft.
-->

## Description
<!-- describe what the PR is about. Explain the approach and possible
drawbacks.It's ok to repeat some text from the related issue. -->

This PR aims to update the `contrast-agnostic` MONAI model to v2.5. This
model should work well on SCI images and whole-spine T1w/T2w images in
addition to previous pathologies and contrasts.

The model architecture (specifically, the number of features in the
network) was updated between v2.4 and v2.5 to account for the increasing
size of the training dataset. Accordingly, the code has been slightly
updated to set the correct number of maximum features in the network
when this model is chosen.

## Testing the PR

1. Check out to the branch `nk/update-contrast-agnostic-to-v2.5`
3. Run `sct_deepseg -task seg_sc_contrast_agnostic -i
<path-to-the-image>.nii.gz` to segment an image (this should
automatically download the latest version replacing the existing one)

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>