`sct_apply_transfo`: Add wrapper function for ANTs binary call
This will allow us to add pre/postprocessing steps that will automatically apply to all
calls to isct_antsApplyTransforms.

I've been wanting to create wrapper functions for a long long time, actually. There's a
failed attempt in a nranch that I've quietly kept alive for years:

https://github.com/spinalcordtoolbox/spinalcordtoolbox/compare/master...jn/3075-refactor-registration-wrappers

Thankfully, because `isct_antsApplyTransforms` is localized to `sct_apply_transfo`,
it should be a lot safer to make this change (vs. the registration-related wrappers
I added in that branch.)