`sct_apply_transfo`: Preserve input datatype if casting is safe
This is pretty expensive, since we have to load both the input and output images. But,
in order to minimize the changes to the ANTs binary, I decided to keep the functionality
as close as possible for my first draft.

Also, there are some built-in numpy cast checking functions (e.g. `can_cast`), but
naturally these fail when checking float64 -> int16. So, I figured it's simplest to
just cast first then compare the arrays.