Fix bug due to discrepancy introduced by `set_qform` and `set_sform` methods (#4697)
## Description

This PR ensures that the qform will always match the sform when
reorienting the image.

Note that with the old method:

- sform == im_dst_aff
- qform ==
[orthogonal(im_dst_aff)](https://github.com/nipy/nibabel/blob/d9c479a03df4119fcac0ffd3bd1b2b0fabf650e7/nibabel/nifti1.py#L1251-L1260)

And because orthogonal(im_dst_affine) may not always equal im_dst_aff,
the two `set_xform` commands may result in different matrices.

Depending on how we utilize the `sform`, this change may result in
numerical differences? So, I am opening this PR also as a way of quickly
checking our test suite.

## Linked issues

Fixes #4689.