Allow `nnunetv2>=2.5.1` due to bugfix for previous issue with `2.4.2` (#4699)
## Description

Previously, we encountered a bug with `nnunetv2==2.4.2` where an empty
segmentation would be produced by our SCI model. We temporarily limited
`nnunetv2==2.3.1`, since that was seemingly the only version that worked
for us. See:

- https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4564

However, there is a newer version of `nnunetv2` (`2.5.1`) that seems to
fix the "empty segmentation" problem.

Note that TotalSpineSeg is incompatible with 2.5.1, so we're probably
going to revert back to 2.3.1 immediately. But, this PR will prepare for
an eventual 2.5.2 release that will finally work for both SCT and
TotalSpineSeg.

## Linked issues

Fixes #4556.