Allow arbitrary footprint shapes in `_dilate_point_labels`
This includes 2D and 3D footprints, and rectangular shapes (rather than just
squares/cubes).

By using `np.maximum.at`, we avoid creating a temporary array just to compute
the maximum. But this is not a significant gain, because the multiplication
in `data_value*footprint[fp_loc]` _does_ create a temporary array.