Use a smaller, equivalent footprint
Inserting a 2D footprint as a single slice of a cube-shape 3D footprint
is very inefficient. For example, with a 5x5 footprint, using a 5x5x5
cube means that the dilation will consider 125 surrounding voxels for
each voxel, and disregard almost all of them (since the footprint is
mostly zero).

Instead, using a 1x5x5 or 5x1x5 or 5x5x1 box means that the dilation
will only consider 25 surrounding voxels.