Use faster dilation algorithm when dilating "mostly zero" point label images (`-x label`) (#4668)
## Description

Since 99.99% of voxels are 0 when `islabel=True`, we don't *need* to run
dilation on every single voxel, which is what skimage will do by
default. Instead, we can fetch the nonzero voxels, extract an area the
size of the footprint, run dilation on that area, and then put it back.

I was concerned that this might behave differently in cases where we
have two labels very close together. But, I think it ends up working
identically, with the larger label dominating the smaller label (thanks
to the use of `maximum` internally in skimage).

## Testing

On my (fast) laptop, dilation time for a typical image went from 7s to
0.15s to <0.01s, thanks to minimizing the area we call
`skimage.dilation` on, and ultimately removing `skimage.dilation`
entirely.

## Linked issues

Fixes #4663.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>