`qc2.py`: Properly index colors for multi-group labels
We assign distinct colormaps for groups of labels (e.g. [1, 2] ... [11, 12, 13, 14]...).
This feature was designed specifically for TotalSpineSeg. However, in our "add_seg_labels"
function, we don't actually take these distinct labels into account. Instead of just creating
a set of unique labels, we need to create the *full range* of labels (including the ones in
between the groups), that way we index the color labels correctly (accounting for gaps).

If we don't index the colors correctly, then the label text will be black instead...