`qc2.py`: Scale the crop radius based on *both* resolutions (H,W)
This is part 1 of the fix for anisotropic images.

Previously, when scaling the crop radius to match the resolution, we only took pixdim[1]
and used it to scale both the height and width of the crop radius. This is wrong! It assumes
that pixdim[1] == pixdim[2] (i.e. isotropic).

So, instead of making that assumption, we take both pixdim[1] and pixdim[2] and scale the
height and width of the crop radius respectively.