`qc2.py`: Apply aspect ratio to `figure` and `imshow`
This is part 2 of the fix for anisotropic images.

The previous commit ensured that we accurately cropped the image for anisotropic images.
However, if we display the mosaic using a 1:1 aspect ratio, the image will appear stretched.
So, we need to scale the height of the figure (as well as the `imshow` call) by the aspect
ratio in order to properly display the image at an isotropic resolution.