`qc2.py`: Fix bug with call to `np.pad` in `mosaic()`
This is one of 2 preexisting bugs that were lurking under the surface.

`np.pad` accepts the argument `pad_width`, which takes the form `((before_1, after_1), ...
 (before_N, after_N))`. However, we pass `radius` directly to `pad_width`, meaning that it
 gets repeated for both axes, and thus the padding is:

 - x axis: (radius[0], radius[1]) i.e. non-symmetrical padding before/after
 - y axis: (radius[0], radius[1]) i.e. non-symmetrical padding before/after

What *should* be happening is this instead:

 - x axis: (radius[0], radius[0]) i.e. symmetrical padding before/after
 - y axis: (radius[1], radius[1]) i.e. symmetrical padding before/after

We never noticed this because for isotropic images, radius[0] == radius[1], and thus the
first case just happened to match the expected behavior.