`qc2.py`: Fix bug due to swapped x/y coords when writing text
This is the second of two preexisting bugs that were lurking under the surface.

In numpy, axis ordering follows zyx convention, instead of the usual (and maybe more intuitive) xyz. (https://stackoverflow.com/a/46857263)

So, when writing text using x/y coordinates, we need to swap the axis order.

Interestingly enough, we got this right in the add_segmentation_labels() function, where
we correctly set (y, x) using the extents[0]/extents[1] dimensions. But, it looks like
we made an error here.

This likely slipped through because radius[0] == radius[1] for isotropic images, and we
usually resample to isotropic.