`sct_maths.py`: Use safe subtraction to avoid int overflow
Using `promote_types` with `np.byte` will result in the smallest possible signed int that a
uint can be safely cast to. If one of the inputs is float, then promote_types will return
float.

Then, using `result_type` will return float > int > uint depending on the two types being
compared. So, if one of the two is float, we'll get a float. Otherwise, we'll get the largest
int that encapsulates both values.

I think this is fine for our purposes?