`qc2.py`: Add binarization step missing in `sct_deepseg` QC
This is done by default for all segmentations in the old `qc.py` QC.

We remembered to add a simplified version of this binarization step when we converted the
`sct_register_multimodal` QC from `qc.py` to `qc2.py`. However, we forgot to add this step
in the `sct_deepseg` QC reports.

Note that this isn't relevant for rootlets/lesion QCs because we don't resample for those.