Add missing binarization+thresholding step for segmentations in `sct_deepseg` QC (#4742)
## Description

Thresholding is done by default (**AFTER** resampling w/linear
interpolation) for all segmentations in the old `qc.py` QC:


https://github.com/spinalcordtoolbox/spinalcordtoolbox/blob/4868a2cdd79a7109bce1de8747f52cbed9d18b0f/spinalcordtoolbox/reports/slice.py#L174-L183

We remembered to add a simplified version of this thresholding step
(that always performs binarization) when we converted the
`sct_register_multimodal` QC from `qc.py` to `qc2.py`:


https://github.com/spinalcordtoolbox/spinalcordtoolbox/blob/f72607f22cdf1debc4e8da148f3827cbfd941667/spinalcordtoolbox/reports/qc2.py#L199-L205

However, we forgot to add this same step in the `sct_deepseg` QC
reports. So, I've added it to both the axial and sagittal mosaic QCs:

Before:


![image](https://github.com/user-attachments/assets/c6b12b5c-9863-46b9-96b0-468a2be62b59)

After:


![image](https://github.com/user-attachments/assets/3121ef5f-d479-4595-bfcb-e288388d96a4)

Note that this isn't relevant for rootlets/totalspineseg QCs because we
don't resample for those.

## Linked issues

Fixes #4741.