`conftest.py`: Improve sorting for better diffs
The reason I even sort by creation time is so that the order of CSV entries roughly matches
the order of commands in the `batch_processing.sh` script. This makes it easier to scan
through diffs and roughly see where the discrepancies first occur.

However, sometimes files are created in a nondeterministic order, nearly simultaneously. So, I
wanted to try rounding the time to the nearest 10s, then sorting by filepath. That way,
for collections of files that are all created in a burst, the order will be a little more
deterministic.

I chose 1s to try to strike a balance between "large enough interval to capture groups of
files" and "not so large that unrelated files get grouped together". 10s felt too large,
and 0.1s felt too small.