Improve how we track changes to SCT's outputs by exporting a summary of all output files (#4759)
## Description

During the lead up to the SCT course, we discovered many subtle behavior
changes that slipped through our testing, and that required manual
examination to catch. Even worse, I almost accidentally introduced a big
change right before the SCT course, and it wasn't caught by the tests
(https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4744#issuecomment-2514633760).
Because of this, I felt especially motivated to try to make some changes
to improve SCT's testing coverage over the holiday break.

IMO, we've got a big gap in our testing. Right now, our tests cover the
following areas:

- **Low sensitivity, high specificity**: Our unit tests for CLI/API are
good at measuring the outputs of specific commands, but often use
measures such as `DICE > 0.95` to avoid spurious failures from small,
acceptable changes over time.
- **High sensitivity, low specificity**: Our integration test
(`batch_processing.sh`) is more sensitive to changes in the processing
pipeline, with small changes to one input often cascading throughout
labeling/registration/CSA steps. However, when it does fail, it can be
hard to pinpoint the exact cause of the change, and so we often blindly
update the cached results because we're limited in how well we can
understand the change.
- We don't really have any **high sensitivity, high specificity** tests
because A) the specificity would make the tests themselves
time-consuming to write and B) the sensitivity would likely cause many
spurious failures, which would again be time-consuming to debug and
update.

To fill this void, I tried writing a test-agnostic function that will
summarize the files in a folder based not only on (size, hash) but also
image properties. That way, we can export the summary as a whole for
`tmp/pytest`, `sct_testing_data`, etc. and compare it between versions
and between OSs.

I'm hoping that diffing these files will provide some insight into _why_
the `batch_processing.sh` output changes, and minimize the time spent
manually comparing images locally, guessing where the change might have
first occurred. And, I'm hoping we can come up with a workflow to diff
the files in a way that won't be disruptive to our PRs.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>