Add new GitHub Actions workflow to test if old releases install without error (#4761)
## Description

This PR is a helper PR, so that I can see where the issues lie with our
old releases. We can then take these fixes (or alternative fixes
depending on the review) and backport them to the corresponding tags.

This is an expensive PR (huge matrix of OSs and versions), so ideally my
hope is that we run this as a cron job maybe... weekly? So that we can
feel safe knowing that older versions still work, without using up a ton
of runners on every PR.

We can also use this workflow to test backported changes before we merge
them. (e.g. if we make a branch off of `5.4` called e.g.
`jn/5.4-backport-fix`, we can also temporarily update this workflow and
add `jn/5.4-backport-fix` to easily install and test that version,
without having to fuss around with installing the older versions
locally.

## Linked issues

Partially addresses #4696.
Partially addresses
https://forum.spinalcordmri.org/t/problem-installing-sct-v5-8-on-linux-mint-21-1/1367.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>