Ensure that `torch<2.6` requirement gets copied over to `requirements-freeze.txt` (#4779)
## Description

This commit is a follow-up to the previous fix of pinning `torch<2.6`
(#4776).

The `grep -v torch` line is preventing the `torch<2.6` fix from being
propagated to the `requirements-freeze.txt` file, causing the Create
Release nightly workflow to fail
(https://github.com/spinalcordtoolbox/spinalcordtoolbox/actions/runs/13050987631).

The `grep -v torch` line was originally added nearly 4 years ago in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3383. At
that time, the state of the repository included several
platform-specific frozen requirements that we wanted to explicitly copy
over, so that we would override the actually-installed version of torch:

https://github.com/spinalcordtoolbox/spinalcordtoolbox/blob/634670e4e0e282d27abaf457def5feb49da9a1f6/requirements.txt#L33-L36

However, we now have a better way of copying over specific lines in
`requirements-freeze.txt` (namely, the string `# append_to_freeze`)

## Linked issues

Fixes #4775.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>