Replace monai-based contrast agnostic model to nnunet-based r20250204 release (#4783)
## Description

This PR aims to update the `contrast-agnostic` model to the nnunet based
one (which was monai-based until now). Experiments showed that the
nnunet model was more robust across difficult subjects than the current
monai-based model. It also fixed issues we were having with shifted
predictions on a few test cases, and works _much more_ better on
compressed cord and whole-spine images.

Notable change is that I set the `keep_largest` argument as the default
for this model as we have seen some cases where the model was predicting
a few pixels outside the cord (quite rarely though).

## Testing the PR

1. Check out to the branch `nk/update-contrast-agnostic-to-v3.1`
2. Run `sct_deepseg -install-task seg_sc_contrast_agnostic`
3. Run `sct_deepseg -task seg_sc_contrast_agnostic -i
<path-to-the-image>.nii.gz` on your test image.

@valosekj or @sandrinebedard could you please checkout to this branch
and try the latest model? Thank you!

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>