Update `courses.rst` with SCT Course 2024 links (#4796)
## Description

This PR updates the documentation with the newest links to the material
and YouTube video (needed to send out final SCT Course email).

## Backporting to 6.5

Now that our documentation links to the latest "stable" release, this
change won't go live immediately. Meaning, we will need to backport this
for it to go live.

We could update the 6.5 tag, but this would require remaking the release
and etc. But because `stable` triggers for the most recent tag, it's
possible to create a secondary tag just for backported documentation
changes. Something as simple as `6.5.docs` would cause `stable` to
update while leaving `6.5` itself unchanged.

## Linked issues

Partially addresses <!--don't close-->#4749.