Port "past release" fixes to the tags themselves, and remove them from the GHA workflow (#4780)
## Description

This PR is part of the plan to backport fixes to the 5.x and 6.x series
of SCT releases. We've already tested the fixes as part of the workflow
using the `perl` find and replace fix. But, now we're going to actually
put those changes into the branches themselves.

I created some temporary "-bp" ("backport") branches based off of
their corresponding tag, that will contain the fixes and allow us to
briefly test them before making any modifications to the tags
themselves. Once we confirm that the fixes work in this new context, we
can update the tags (and the releases), and revert this PR back to
testing the tags.

## Linked issues

Fixes #4777. 
Fixes #4696. (When tags are updated.)
Fixes #4670. (When tags are updated.)

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>