Allow adding `-qc-seg` to crop the QC generated by `sct_deepseg` (#4754)
## Description

I am adding the possibility to pass an existing segmentation to be used
as a cropping box in the QC generated by `sct_deepseg`.
Below are detailed the choices made for both QC: 

For the axial QC:
- S-I resolution is same as the input image
- the viewed images are resampled to 0.6mm
- if a segmentation is given for QC
- only the slices segmented (+3 before and after the segmentation) are
kept
- the centers of the cropped image are centered on the given QC
segmentation
- for the radius: the width and height of the images are the maximum
between the max width/height (increased by 20%) of the non-empty regions
of the QC segmentation and (15,15)

For the sagittal QC: 
- R-L resolution is cropped to have only 30 slices if it has more than
30 slices and if not QC segmentation is given or kept the same
- the viewed images are resampled to 0.6 mm
- if a QC segmentation is given: 
  - only the segmented slices (+2 before and after) are kept
- the centers are computed based on the centers of mass of each slice of
the QC segmentation
- for the radius: the width is chosen as maximum between 1/4 of the
image width and the maximum width of the segmentation. The height is
computed to keep the full I-S view (i.e. half the image height).

## Linked issues

Fixes
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4705.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>
Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>