`install_sct`: Remove outdated `force_bashrc_loading`
This function is deeply invasive, and may result in unexpected behavior on macOS systems
that explicitly don't wish to use `.bashrc`.

Perhaps this was needed at some point (prior to the macOS switch to zsh), but nowadays,
macOS machines don't even use `.bashrc`. So, this was doing less than nothing for SCT
installations!