Update web tutorials to match the changes made for the 2024 SCT Course (#4804)
## Description

This PR mimics the following PR: 

- https://github.com/spinalcordtoolbox/sct_tutorial_data/pull/25

The above PR contains all of the changes we made to the
`sct_tutorial_data` repo to get it to match the slides. Because of the
fact that `sct_tutorial_data` was previously in sync with the web
tutorials, I figured that any updates to that repo would just need to be
applied to the web tutorials on a 1:1 basis to get everything back in
sync.

Therefore, the commits in this PR directly reference other commits.
However, a handful of `sct_tutorial_data` commits were not applicable:

## Linked issues

Fixes #4749.
Fixes #4799.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>
Co-authored-by: Jan Valosek <39456460+valosekj@users.noreply.github.com>
Co-authored-by: sandrine <sandrine.bedard@polymtl.ca>