`conftest.py`: Exclude `straightening.cache` from summaries
This file contains the `__version__` string in plaintext, meaning that it will never
match between `master` and the PR. It also doesn't contain any useful numerical
information, so it's not worth it to try to preserve it like we do the .csv files.

Maybe it's still worth noting whether this file is getting generated at all? But,
I don't think it's critical for the purpose of the "file summary" check, which
is to catch minute numerical differences with PRs.