Compare new `batch_processing.sh` file summaries between PRs and `master` branch (#4803)
## Description

This PR is a follow-up to:

- https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4759

In that PR, because the workflow hadn't been merged yet, comparisons
could only be done between platforms. But, now that it's been running on
`master` for a few weeks, we can compare the results of PRs against the
cached results from master.

## Linked issues

Follow-up to #4759.